Split concept setup into 3 modules
One to download concepts, one to create concepts from the downloaded list, and one to call the aforementioned modules at runtime. Updated tests to work with new multi module setup
